>Combined all the home-component (navigation bar, search bar, catergory bar) into the HomeView.vue file. >created separate folder for css (assets folder) >moved images folder to assets >currently debugging for an view error(when many rooms we can't scroll down).
