+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <link rel="icon" 
     type="image/webp" 
-    href="/src/images/logo/logo.webp">
+    href="src\assets\images\logo\transparentlogo 1080.png">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>ViCoSpaces</title>
     <style>
@@ -17,8 +17,7 @@
         padding: 0;
         font-family: Arial, Helvetica, sans-serif;
         position: fixed;
-        overflow-y: scroll;
-        height: 100%;
+        overflow-y: auto;
         width: 100%;
 
       }
