+++ src/assets/styles/homeView.css
@@ -1,0 +1,262 @@
+
+  main{
+    background-color: palevioletred;
+    height: 100vh;
+    flex-grow: 1;
+  }
+  .room-view-container {
+    position: relative;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    background-color: red;
+    justify-content: first baseline;
+    padding: 20px;
+    
+    width: 100%;    
+  }
+  .room-form {
+    width: 100%;
+    height: auto;
+    display: grid;
+    grid-auto-flow: row;
+    grid-template-columns: repeat(3, minmax(370px, 1fr));
+    grid-template-rows: auto; 
+    gap: 30px;
+    flex-grow: 1;
+    overflow-y: auto;
+  }
+  
+  .room {
+    position: relative; 
+    border-radius: 12px;
+    height: 350px;
+    width: 100%;
+    box-sizing: border-box;
+    overflow: hidden;
+    word-wrap: break-word;
+    box-shadow: 3px 5px 20px #9e9e9e;
+  }
+  
+  .image-content {
+    position: relative;
+    background-image: url(../images/SVG/room-background_2.svg);
+    display: grid;
+    grid-template-columns: 28% 1fr 1fr;
+    grid-template-rows: auto auto;
+    border-radius: 12px 12px 0px 0px;
+    height: 60%;
+    box-sizing: border-box;
+    background-color: #2d8eff;
+  }
+  .hosting-container {
+    grid-column: 3;
+    background-color: rgb(255, 255, 255);
+    height: 64px;
+    width: 64px;
+    border-radius: 50%;
+    align-self: center;
+    transform: translate(107px, -20px);
+    background-image: url("../images/UserProfileStocks/megan.jpg");
+    background-size: 220%;
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
+  }
+  .text-content {
+    display: flex;
+    flex-direction: column;
+    height: 40%;
+    max-height: 100%;
+    max-width: 100%;
+    border-radius: 0px 0px 12px 12px;
+    background-color: rgb(255, 255, 255);
+    padding: 0px 20px 10px 20px;
+    align-items: center;
+    box-sizing: border-box;
+  
+  }.room-name-container{
+    width: 100%;
+    display: flex;
+    justify-content: flex-start;
+  }
+  .joined-and-button-container{
+    display: flex;
+    justify-content: space-between;
+    align-content: space-between;
+    width: 100%;
+    height: 80%;
+    
+  }
+  .joined-users{
+    display: flex;
+    align-self: center;
+    height: 50%;
+    width: 30%;
+  }
+  .user-wrapper{
+    width: 100%;
+    display: flex;
+  
+  }
+  .user-1-container{
+      width: 60px;
+      height: 60px;
+      border-radius: 50%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      color: white;
+      background-color: greenyellow;
+      font-size: 24px;
+      font-weight: bold;
+      position: relative; /* Allows us to position relative to the container */
+      margin-left: -20px;
+  }
+  .user-1{
+    width: 32px;
+    height: 32px;
+    border: 1px solid #2d8eff;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+    position: relative; /* Allows us to position relative to the container */
+    margin-left: -1px;
+    
+  }
+  .user-2{
+    width: 32px;
+    height: 32px;
+    border: 1px solid #2d8eff;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+    position: relative; /* Allows us to position relative to the container */
+    margin-left: -9px;
+  }
+  .user-3{
+    width: 32px;
+    height: 32px;
+    border: 1px solid #2d8eff;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+    position: relative; /* Allows us to position relative to the container */
+    margin-left: -9px;
+  }
+  .user-count{
+    width: 32px;
+    height: 32px;
+    border: 1px solid #2d8eff;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: #2d8eff;
+    color: white;
+    font-size: 24px;
+    font-weight: bold;
+    position: relative; /* Allows us to position relative to the container */
+    margin-left: -9px;
+  }
+  .number-count{
+    font-size: 0.9rem;
+    
+  }
+  
+  .join-btn-container{
+    align-content: center;
+    height: 100%;
+  }
+  .join-btn {
+    font-weight: bolder;
+    border: none;
+    border-radius: 12px;
+    height: 42px;
+    width: 124px;
+    color: white;
+    background-color: #2d8eff;
+  }
+  button {
+    cursor: pointer;
+  }
+  @media (max-width: 1386px) {
+    main {
+      height: 100vh;
+      padding: 20px 38px 0px 38px;
+      overflow-y: auto ;
+    
+    }
+    .image-content {
+    position: relative;
+    background-image: url(../images/SVG/room-background_2.svg);
+    display: grid;
+    grid-template-columns: 28% 1fr 1fr;
+    grid-template-rows: auto auto;
+    border-radius: 12px 12px 0px 0px;
+    height: 60%;
+    background-color: #2d8eff;
+  }
+  
+  .hosting-container {
+    position: absolute;
+    top: 35px;
+    right: 120px;
+    background-color: rgb(255, 255, 255);
+    height: 64px;
+    width: 64px;
+    border-radius: 50%;
+    background-image: url("../images/UserProfileStocks/megan.jpg");
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
+  }
+  
+  }
+  
+  @media (max-width: 1228px) {
+    .room-form {
+      grid-template-columns: repeat(2, 1fr); /* Two columns for smaller screens */
+    }
+  }
+  
+  @media (max-width: 820px) {
+  }
+  @media (max-width: 794px){
+    main {
+      height: 100vh;
+      padding: 20px 20px 0px 20px;
+      overflow-y: auto ;
+    
+    }
+    .room-form {
+      grid-template-columns: 1fr; /* One column for even smaller screens */
+    }
+  }
+  
+  @media (max-width: 600px) {
+    .room-form {
+      grid-template-columns: 1fr; /* One column for even smaller screens */
+    }
+    .room-view-container{
+      background-color: #9e9e9e;
+    }
+  }
+  @media (max-width: 375px) {
+    main{
+      padding: 20px;
+    }
+  
+  }
