+++ src/router/index.js
@@ -9,7 +9,7 @@ const router = createRouter({
     { path: '/', component: () => import('../views/LandingPageView.vue'), meta: { requiresAuth: false }  },
     { path: '/about', component: () => import('../views/AboutView.vue'), meta: { requiresAuth: false }  },
     { path: '/roadmap', component: () => import('../views/RoadMap.vue'), meta: { requiresAuth: false }  },
-    { path: '/home', component: () => import('../views/HomeView.vue'), meta: { requiresAuth: true } },
+    { path: '/home', component: () => import('../views/HomeView.vue'), meta: { requiresAuth: false } },
     { path: '/register', component: () => import('../views/RegisterView.vue'), meta: { requiresAuth: false } },
     { path: '/login', component: () => import('../views/SignInView.vue'), meta: { requiresAuth: false } },
     {
